requirements.txt
pandas==2.0
psycopg2-binary==2.9.6
jupyterlab==4.0
scikit-learn==1.2
bokeh==3.1